Restructure of the GUI to utilize threading
This also makes the code a LOT more stable. The PID controller calculations
within an ISR were causing the microcontroller to periodicially crash, as the ISRs
were taking too long to complete and were backing up. Instead, a flag is now set
so that the background processing can occur without blocking interrupts.

+++ Microcontroller/AVR/src/PID.c
@@ -1,8 +1,8 @@
 #include "BeerPong.h"
 
-double updatePID(PID *controller, double error, double position)
+float updatePID(PID *controller, float error, float position)
 {
-	double pTerm, iTerm, dTerm;		
+	float pTerm, iTerm, dTerm;		
 	
 	//Calculate the P term
 	pTerm = controller->pGain * error;
